\name{age.dud}
\alias{age.dud}
\docType{data}
\title{ age of sediment samples from a sediment core taken of lake Dudinghausen}
\description{
  age of the sediment samples within the data set dud.df
}
\usage{data(age.dud)}

\references{
Dressler, M., Selig, U., Doerfler, W., Adler, S., Schubert, H., Huebener, T. 2006. Environmental changes and the migration period in Europe by the example of Lake Dudinghausen northern Germany. Quat. Res. 66:25-37. 

}
\examples{
data(age.dud)
}
\keyword{datasets}
