\name{age_dud}
\docType{data}
\alias{age_dud}

\title{
age of sediment samples from the core of lake Dudinghausen
}

\description{
Age before present of the core samples of lake Dudinghausen
}

\usage{data(age_dud)}

\format{A vector containing 133 values.}

\details{Lake Dudinghausen is a recent eutrophic lake.  
 The lake morphology of lake Dudinghausener (DUD) (53.54.30 N, 12.12.40 E)
 is characterized by a nearly homogeneously shaped basin. The lake is almost completely 
 surrounded in agricultural used watershed land, and only a small part of the northern shore is forested. 
 Large areas of cattail are located at the western and eastern edges. The age is given as age before present.
}


\references{
 Dressler, M., Selig, U., Doerfler, W., Adler, S., Schubert, H., Huebener, T. 2006.
 Environmental changes and the migration period in Europe by the example of Lake Dudinghausen 
 northern Germany. Quat. Res. 66:25-37.
}

\keyword{datasets}
