\name{train_set.MV}
\alias{train_set.MV}
\docType{data}
\title{North German diatom training set}
\description{
The Mecklenburg-Vorpommern data set consists of 84 lakes located in the NE German lowlands
(Mecklenburg-Vorpommern). The lakes are typical kettle-type, formed in glacial till 
of Weichselian origin, and are well-buffered, hardwater and carbonate-rich lakes. 
Intensification of agricultural in the lake catchments during recent decades has resulted
in more or less eutrophic conditions in the most of the lakes. 
}
\usage{data(train_set.MV)}
\format{
 data.frame
}

\source{www.biologie.uni-rostock.de/abt/botanik/AG-Phykologie/index-engl.htm
}
\references{Adler,S., Huebener T.,Anderson J.N., Lotter,A.F.,Modelling the distribution of diatoms along contrasting
	environmental gradients in Europe using GLM, WA and Proportional Odds Models, submitted in Journal of European 
	Phycology.}
\examples{
data(train_set.MV)

}
\keyword{datasets}
