\name{train_env.MV}
\alias{train_env.MV}
\docType{data}
\title{ North German diatom training set - Total Phosphorous}
\description{
  For each lake, water samples were collected three times a year (spring, summer and autumn) at a water depth of 0.5 m and analysed for 19 environmental parameters using standard methods. All lakes, excepting nine shallow lakes are dimictic and have surface areas >50 ha. The here given Total Phosphorous (TP) values are the mean from spring, summer and autumn TP. The data are log10 transformed. 

}
\usage{data(train_env.MV)}
\format{A data frame with 84 observations }

\source{www.biologie.uni-rostock.de/abt/botanik/AG-Phykologie/index-engl.htm }
\author{Sven Adler and Thomas Huebener}
\references{in preperation}

\examples{
data(train_env.MV)
}
\keyword{datasets}
