\name{train_set.MV}
\alias{train_set.MV}
\docType{data}
\title{ North German diatom training set
}
\description{
  The Mecklenburg-Vorpommern data set consists of 84 lakes located in the NE German lowlands (Mecklenburg-Western Pomerania, Mecklenburg-Vorpommern (MV)). The lakes are typical kettle-type, formed in glacial till of Weichselian origin, and are well-buffered, hard water and carbonate-rich lakes. Intensification of agricultural in the lake catchments during recent decades has resulted in more or less eutrophic conditions in the most of the lakes. Species codes are following the EDDI code http://craticula.ncl.ac.uk/Eddi/jsp/index.jsp 
}
\usage{data(train_set.MV)}

\source{
  www.biologie.uni-rostock.de/abt/botanik/AG-Phykologie/index-engl.htm 
}
\author{Sven Adler and Thomas Huebener}
\examples{data(train_set.MV)}
\keyword{datasets}
